tests(helpers): separate miscellaneous and shell functions (#13542)
It is one of the serial refactors of helpers.lua
KAG-5236

The function `kong_exec()` needs an upvalue `conf`, which must be reloaded each time.
